fix: move to cusom download handler
